Introduce new syntax for setting the parameter names for blocks
The block header will look like the following:

<optional global id> <optional parameter name ":"><block type>

This is useful when the same block type can be set to multiple parameters,
or the parameter is an interface type.

E.g. the `Array` block (schema's Array) has the `Items` parameter which is a Schema interface type.
To define an array of integers as a directive, you would write:

```
@array {
  items:string {}
}
```