Generate the JSON representation for an object's JSON schema correctly
The schema.Object's Properties parameter has the Conflow parameter names as
keys, but the JSON property names should be used when JSON
marshalling/unmarshalling.

For this purpose this commit introduces the following changes:
 - add fieldNames map to the object schema: json property name -> field name mapping
 - add parameterNames map to the object schema: json property name -> parameter name mapping

Our schema.Object will still contain the data in a format that's useful for us,
so some conversion will happen in MarshalJSON()/UnmarshalJSON().